ALB-42 - Fix compile warnings
Because some pointers were declared with u32 size, warnings were issued.
The types were replaced with uintptr_t type, which determines the size
based on system settings. Other errors were caused by non-void functions
which didn't return any value.

+++ include/fsl_esdhc.h
@@ -171,9 +171,9 @@ struct fsl_esdhc_cfg {
 #ifdef CONFIG_FSL_LAYERSCAPE
 	u64	esdhc_base;
 #else
-	u32	esdhc_base;
+	uintptr_t esdhc_base;
 #endif
-	u32	sdhc_clk;
+	uintptr_t sdhc_clk;
 	u8	max_bus_width;
 	struct mmc_config cfg;
 };
